refactor: 주석 추가
